KBV-179 Restructure project layout to have a root level lib and lambdas folder.
This makes the project structure similar to di-ipv-core-back and di-ipv-cri-address-api
