--- build.gradle
@@ -1,87 +1,0 @@
-plugins {
-	id 'idea'
-	id 'application'
-	id 'java'
-	id "jacoco"
-	id "com.diffplug.spotless" version "6.0.4"
-	id "io.swagger.core.v3.swagger-gradle-plugin" version "2.1.11"
-	id "org.sonarqube" version "3.3"
-	id "io.freefair.aspectj.post-compile-weaving" version "6.3.0"
-}
-
-group "uk.gov.di.ipv.cri.kbv.api"
-
-repositories {
-	mavenCentral()
-}
-
-sourceCompatibility = 11
-
-dependencies {
-	implementation "com.amazonaws:aws-lambda-java-core:1.2.1",
-			'ch.qos.logback:logback-classic:1.2.10',
-			'ch.qos.logback:logback-access:1.2.10',
-			"software.amazon.awssdk:url-connection-client:2.17.121",
-			"com.amazonaws:aws-lambda-java-events:3.11.0",
-			'com.fasterxml.jackson.core:jackson-databind:2.13.1',
-			'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.13.1',
-			'com.amazonaws:aws-java-sdk-dynamodb:1.12.150',
-			'software.amazon.awssdk:dynamodb-enhanced:2.17.121',
-			"org.hibernate.validator:hibernate-validator:7.0.1.Final",
-			"org.hibernate:hibernate-validator-annotation-processor:7.0.1.Final",
-			"com.fasterxml.jackson.datatype:jackson-datatype-jsr310",
-			"javax.validation:validation-api:2.0.1.Final",
-			"javax.el:javax.el-api:3.0.0",
-			"org.glassfish:javax.el:3.0.0",
-			'com.nimbusds:nimbus-jose-jwt:9.18'
-
-	aspect "software.amazon.lambda:powertools-tracing:1.10.2"
-
-	testImplementation "org.junit.jupiter:junit-jupiter:5.8.2",
-			'org.mockito:mockito-junit-jupiter:4.3.1',
-			'org.mockito:mockito-inline:4.3.1'
-}
-
-test {
-	useJUnitPlatform ()
-	finalizedBy jacocoTestReport
-}
-
-jacocoTestReport {
-	dependsOn test
-	reports {
-		xml.enabled true
-	}
-}
-
-spotless {
-	java {
-		target "src/main/**/*.java"
-		googleJavaFormat("1.13.0").aosp()
-		importOrder "", "javax", "java", "\\#"
-		endWithNewline()
-	}
-	groovyGradle {
-		target '**/*.gradle'
-		greclipse()
-		trimTrailingWhitespace()
-		endWithNewline()
-	}
-}
-
-run {
-	debugOptions {
-		enabled = true
-		port = 8082
-		server = true
-		suspend = false
-	}
-}
-
-sonarqube {
-	properties {
-		property "sonar.projectKey", "alphagov_di-ipv-cri-kbv-api"
-		property "sonar.organization", "alphagov"
-		property "sonar.host.url", "https://sonarcloud.io"
-	}
-}
