+++ lambdas/session/build.gradle
@@ -1,0 +1,64 @@
+plugins {
+	id "java"
+	id "application"
+	id "idea"
+	id "jacoco"
+	id "io.freefair.aspectj.post-compile-weaving" version "6.3.0"
+}
+
+repositories {
+	mavenCentral()
+}
+
+dependencies {
+	implementation project(":lib"),
+			"com.amazonaws:aws-lambda-java-core:1.2.1",
+			"ch.qos.logback:logback-classic:1.2.10",
+			"ch.qos.logback:logback-access:1.2.10",
+			"software.amazon.awssdk:url-connection-client:2.17.121",
+			"com.amazonaws:aws-lambda-java-events:3.11.0",
+			"com.fasterxml.jackson.core:jackson-databind:2.13.1",
+			"com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.13.1",
+			"com.amazonaws:aws-java-sdk-dynamodb:1.12.150",
+			"software.amazon.awssdk:dynamodb-enhanced:2.17.121",
+			"org.hibernate.validator:hibernate-validator:7.0.1.Final",
+			"org.hibernate:hibernate-validator-annotation-processor:7.0.1.Final",
+			"com.fasterxml.jackson.datatype:jackson-datatype-jsr310",
+			"javax.validation:validation-api:2.0.1.Final",
+			"javax.el:javax.el-api:3.0.0",
+			"org.glassfish:javax.el:3.0.0",
+			"com.nimbusds:nimbus-jose-jwt:9.18"
+
+	aspect "software.amazon.lambda:powertools-tracing:1.10.2"
+
+	testImplementation "org.junit.jupiter:junit-jupiter:5.8.2",
+			"org.mockito:mockito-junit-jupiter:4.3.1",
+			"org.mockito:mockito-inline:4.3.1"
+
+}
+
+java {
+	sourceCompatibility = JavaVersion.VERSION_11
+	targetCompatibility = JavaVersion.VERSION_11
+}
+
+task buildZip(type: Zip) {
+	from compileJava
+	from processResources
+	destinationDirectory = file("$rootDir/dist")
+	into("lib") {
+		from configurations.runtimeClasspath
+	}
+}
+
+test {
+	useJUnitPlatform ()
+	finalizedBy jacocoTestReport
+}
+
+jacocoTestReport {
+	dependsOn test
+	reports {
+		xml.required.set(true)
+	}
+}
