+++ lambdas/session/src/test/java/uk/gov/di/ipv/cri/kbv/api/handler/TestData.java
@@ -1,0 +1,18 @@
+package uk.gov.di.ipv.cri.kbv.api.handler;
+
+public final class TestData {
+    public static final String APERSONIDENTITY = "{\n" +
+            "      \"dateOfBirth\": \"0200-01-10\",\n" +
+            "      \"surname\": \"JOE\",\n" +
+            "      \"firstName\": \"BLOGGS\",\n" +
+            "      \"addresses\": [\n" +
+            "        {\n" +
+            "          \"addressType\": \"CURRENT\",\n" +
+            "          \"houseNumber\": \"1\",\n" +
+            "          \"townCity\": \"EURASIA\",\n" +
+            "          \"street\": \"LONG BEFORE CLIMATE CHANGE\",\n" +
+            "          \"postcode\": \"THE CAVE\"\n" +
+            "        }\n" +
+            "      ]\n" +
+            "    }";;
+}
