+++ lib/build.gradle
@@ -1,0 +1,64 @@
+plugins {
+	// Apply the java-library plugin for API and implementation separation.
+	id "java-library"
+	id "idea"
+	id "jacoco"
+}
+
+group "uk.gov.di.ipv.cri.kbv.api"
+
+repositories {
+	// Use Maven Central for resolving dependencies.
+	mavenCentral()
+}
+
+dependencies {
+
+	implementation "com.amazonaws:aws-lambda-java-core:1.2.1",
+			"com.amazonaws:aws-lambda-java-events:3.11.0",
+			"com.amazonaws:aws-java-sdk-dynamodb:1.12.122",
+			"com.nimbusds:oauth2-oidc-sdk:9.3.1",
+			"com.fasterxml.jackson.core:jackson-core:2.13.0",
+			"com.fasterxml.jackson.core:jackson-databind:2.13.0",
+			"com.fasterxml.jackson.core:jackson-annotations:2.13.0",
+			"com.google.code.gson:gson:2.8.9",
+			"org.apache.httpcomponents:httpclient:4.5.13",
+			"org.apache.commons:commons-lang3:3.5",
+			"org.slf4j:slf4j-simple:1.7.32",
+			"software.amazon.awssdk:dynamodb-enhanced:2.17.89",
+			"com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:2.13.0",
+			"software.amazon.lambda:powertools-parameters:1.8.0",
+			"javax.validation:validation-api:2.0.1.Final"
+
+	testImplementation "org.junit.jupiter:junit-jupiter:5.8.2",
+			"org.mockito:mockito-core:4.1.0",
+			"org.mockito:mockito-junit-jupiter:4.1.0",
+			"com.github.tomakehurst:wiremock-jre8:2.31.0",
+			"uk.org.webcompere:system-stubs-jupiter:1.1.0",
+			"com.fasterxml.jackson.datatype:jackson-datatype-jsr310"
+}
+
+java {
+	sourceCompatibility = JavaVersion.VERSION_11
+	targetCompatibility = JavaVersion.VERSION_11
+	withSourcesJar()
+}
+
+tasks.named("jar") {
+	manifest {
+		attributes("Implementation-Title": project.name,
+		"Implementation-Version": project.version)
+	}
+}
+
+test {
+	useJUnitPlatform ()
+	finalizedBy jacocoTestReport
+}
+
+jacocoTestReport {
+	dependsOn test
+	reports {
+		xml.required.set(true)
+	}
+}
