+++ settings.gradle
@@ -1,0 +1,4 @@
+rootProject.name = "di-ipv-cri-kbv-api"
+include "lib",
+		"lambdas",
+		"lambdas:session"
